Add Travis-CI YML file, Sample Test script
Only contains minimal info - dependencies might not be installed on
Travis machine - yet to determine. If not, they can be run by the
"install" part, or installed locally.
Also contains a minimal test script which should return true. Construct
a test script of your own and add it to this as required.